+++ CHANGELOG.md
@@ -13,6 +13,7 @@
   - Change HTML parameters from countryID and topic to id and report ([#30])
   - Mention API and GitHub on About page ([#5])
   - On Click to marker now only zooms to polygon instead to fixed value 
+- Return a GeoJSON when computing an indicator from the CLI using a dataset and FID ([#57])
 
 [#28]: https://github.com/GIScience/ohsome-quality-analyst/pull/28
 [#40]: https://github.com/GIScience/ohsome-quality-analyst/pull/40
@@ -25,6 +26,7 @@
 [#30]: https://github.com/GIScience/ohsome-quality-analyst/issues/30
 [#5]: https://github.com/GIScience/ohsome-quality-analyst/issues/5
 [#61]: https://github.com/GIScience/ohsome-quality-analyst/pull/61
+[#57]: https://github.com/GIScience/ohsome-quality-analyst/pull/57
 
 
 ## 0.3.1